Se arregla error al validar permisos en userproperties en organization, se cambia consulta apra validar en el middleware

+++ routes/api.php
@@ -157,9 +157,9 @@ Route::get('organizations/{organization}/userproperties/{id}', 'OrganizationUser
 Route::group(
     ['middleware' => 'auth:token'],
     function () {
-        Route::post('organizations/{organization}/userproperties', 'OrganizationUserPropertiesController@store');
-        Route::put('organizations/{organization}/userproperties/{id}', 'OrganizationUserPropertiesController@update');
-        Route::delete('organizations/{organization}/userproperties/{id}', 'OrganizationUserPropertiesController@destroy');
+        Route::post('organizations/{organization}/userproperties', 'OrganizationUserPropertiesController@store')->middleware('permission:create');
+        Route::put('organizations/{organization}/userproperties/{id}', 'OrganizationUserPropertiesController@update')->middleware('permission:update');
+        Route::delete('organizations/{organization}/userproperties/{id}', 'OrganizationUserPropertiesController@destroy')->middleware('permission:destroy');
     }
 );
